adds the md2-switch for number_section
By default the `number_section` argument is set to false if used with `bookdown::markdown_document2`.
In case the auto-numbering feature of bookdown in this combination of bookdown and officedown is needed the new md2-switch is added. Its default value is `FALSE`.
I found no native way to check, if `number_section` is set to `FALSE` manually or by code. Hence a new parameter seems to be a straightforward solution.